wip: reorganized code to better handle dependencies
