+++ flake.nix
@@ -32,7 +32,7 @@
         in
         {
           go-hello = pkgs.buildGoModule {
-            pname = "orchestra-go";
+            pname = "gitsaver";
             inherit version;
             # In 'nix develop', we don't need a copy of the source tree
             # in the Nix store.
